moving files for more training + updating algo and dataset
