--- Data/TempTrainingData.txt
@@ -1,2 +1,0 @@
-2025,08,12,00,20,20,100,0,,140,5.4,,1014.2,,5
-2025,08,12,00,20,20,100,0,,140,5.4,,1014.2,,5
