+++ Data/weather_with_future.csv
@@ -190,4 +190,3 @@
 2025,08,19,20,27.2,21.7,72,0,,30,7.6,,1016.9,,2,25
 2025,08,19,21,25,22.1,84,0,,310,5.4,,1017.6,,2,22.8
 2025,08,19,22,22.8,21.8,94,0,,0,0,,1018.4,,1,22.2
-2025,08,19,23,22.2,21.7,97,0,,0,0,,10185,,
