+++ Netowrk_attempt_1/ReadMe.md
@@ -1,0 +1,2 @@
+This attempt I used all the data and created random networks
+I ran the networks semi async, and tracked the info in output0 and output1
