+++ Netowrk_attempt_1/model1.txt
