+++ Netowrk_attempt_1/model10-6.txt
