+++ Netowrk_attempt_1/model10-8-3.txt
