+++ Netowrk_attempt_1/model10-8-6.txt
