+++ Netowrk_attempt_1/model8-4.txt
