+++ Netowrk_attempt_1/model8-7-3.txt
