+++ Netowrk_attempt_1/model8-8.txt
