+++ Netowrk_attempt_1/output0.txt
