+++ Netowrk_attempt_1/output1.txt
