+++ Network_attempt_2/model1.txt
@@ -1,0 +1,2 @@
+n n 
+n665.9969050288748_-535.7144650481382_474.5245948237383
