+++ src/NetworkUtil.java
@@ -1,4 +1,4 @@
-package FirstExamples;
+package src;
 
 import java.io.File;
 import java.io.FileWriter;
