+++ src/createNetworks.java
@@ -1,4 +1,4 @@
-package FirstExamples;
+package src;
 
 // This is used to create new networks with minor edits
 public class createNetworks {
@@ -7,7 +7,7 @@ public class createNetworks {
     public static void main(String[] args) {
         int[] colArr = {1};
     
-        new NetworkTrainer("TempuratureBiases/model0.txt", "Data/weather_with_future.csv", colArr);
+        new NetworkTrainer("Network_attempt_2/model1.txt", "Data/test.csv", colArr);
     
     }
